Change the way configs are handeled
