+++ docs/source/api/derotation.config.load_config.rst
@@ -1,0 +1,12 @@
+﻿derotation.config.load\_config
+==============================
+
+.. automodule:: derotation.config.load_config
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      load_config
+      update_config_paths
