+++ docs/source/api/derotation.derotate_batch.rst
@@ -3,11 +3,9 @@
 
 .. automodule:: derotation.derotate_batch
 
-   
+
    .. rubric:: Functions
 
    .. autosummary::
-   
+
       derotate
-      update_config_paths
-
